refactor: a -> BoxofficeInfo 클래스명 변경, decodeData 메서드 구현
