+++ BoxOffice/BoxOfficeInfo.plist
@@ -3,6 +3,6 @@
 <plist version="1.0">
 <dict>
 	<key>APIKEY</key>
-	<string>APIKEY 입력</string>
+	<string>84be56724fec6406422449cf1f43afc2</string>
 </dict>
 </plist>
